+++ src/app/download_control.rs
@@ -14,8 +14,8 @@ impl PydlApp {
             f(&mut core);
         }
         {
-            let core = shared.lock();
-            core_sync::sync_core_to_app(&core, self);
+            let mut core = shared.lock();
+            core_sync::sync_core_to_app(&mut core, self);
         }
     }